#293 BaseYamlNode.toString() and YamlPrintException
